requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
requests==2.31.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.3.0
python-dotenv==1.0.0